# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gestion Locative API
  description: |
    API REST pour l'application SaaS de gestion locative.
    
    ## Authentication
    L'API utilise JWT Bearer tokens. Incluez le header `Authorization: Bearer <token>` dans vos requêtes.
    
    ## Idempotency
    Pour les opérations POST qui modifient l'état (paiements, signatures), incluez le header `Idempotency-Key: <uuid>`.
    
    ## Rôles
    - **admin**: Accès complet
    - **owner**: Gestion de ses propriétés
    - **tenant**: Gestion de son bail et paiements
    - **provider**: Gestion de ses interventions
  version: 1.0.0
  contact:
    name: Support
    email: support@gestionlocative.fr
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /properties/{pid}:
    get:
      tags: [Properties]
      summary: Détails d'une propriété
      security:
        - bearerAuth: []
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails de la propriété
        '404':
          description: Propriété non trouvée
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...